backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
slowapi
python-dotenv
pyjwt
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pytest
pytest-asyncio
httpx